[XLA:GPU] Extend `BlockLevelParameters` and simplify API to Triton IR emitter.
`BlockLevelParameters` now contains also `num_warps`, `num_stages`, and
`num_ctas`, which are parameters shared across the new and legacy
infrastructures. By pushing the logic to decide which Triton IR emitter to use
deeper into the stack, it becomes possible to produce a saner signature for
`TritonWrapper`, `Emit{Generic,MatMul,SoftMax}`, `CreateTritonModule`, and
`CompileTritonToLLVM`.

A side benefit is that the legacy infrastructure for matrix multiplications
now always parses its tiling configuration from the `backend_config` of the
relevant fusion instruction. This uncovered the potential for mismatches
(and an actual mismatch in a test), where the provided `TritonGemmConfig` does
not correspond to the `TritonGemmKey` embedded in the fusion instruction's
backend config. It also enables us to more easily control which path a
SoftMax fusion is expected to take---between the legacy SoftMax emitter and
the new generic emitter.

The next steps will be to

1. encode the configuration for generic Triton fusions in the backend config
   of the fusion instruction;
2. explicitly redirect as many Triton SoftMax fusions towards the new generic
   emitter as is possible;
3. delete the specialized Triton SoftMax emitter in favour of the generic
   emitter as soon as all the fusions in 2. have been successfully converted.

PiperOrigin-RevId: 642612991