[XLA:MSA] Sync copy replacement
This CL implements the sync copy replacement feature in MSA. The feature is disabled by default for now.

The feature is implemented by creating a closure of intervals for each buffer in sorted_buffer_intervals. Each closure has an entrance interval and all intervals of a closure are connected to the entrance interval through at least one sync copy instruction that we are trying to replace. The closure is then processed in a loop, and the allocation values for the closure are created. If all allocation values of the closure can be allocated successfully, and the sync copy instruction replacement correctness conditions are met, the sync copy will be replaced with an async copy instruction. If not, the sync copy instruction is kept as is, and will be ignored in replacement efforts of future closures. For a sync copy to be consider replaceable, beside not being in the ignored list because of the past failures, it must not change the layout between its input and output, it must not have a preassigned memory space for its operand or output, and its operand value must only show up in one position.

The CL also includes a number of changes to the MSA algorithm to support the sync copy replacement feature. These changes include:

* Adding a new field to the AllocationRequest struct to indicate whether the request requires a copy allocation.
* Adding a new field to the AllocationRequest struct to store the latest time that the copy allocation is required.
* Adding a new field to the AllocationRequest struct to store the instruction that the copy allocation is required for.
* Adding a new method to the MsaAlgorithm class to get the corrected use time for an instruction.
* Adding a new method to the MsaAlgorithm class to create the allocation values for a closure of intervals.
* Adding a new method to the MsaAlgorithm class to check if an instruction is in the set of pending copy replacements.
* Adding a new method to the MsaAlgorithm class to clear the pending chunks and asynchronous copies.
* Adding a new field to the MsaAlgorithm class to store the set of processed values.
* Adding a new field to the MsaAlgorithm class to store the set of failed copy replacements.

The CL also includes a number of changes to the MSA algorithm to improve the performance of the sync copy replacement feature. These changes include:

* Using a worklist to keep track of the uses that need to be processed.
* Using a hash map to store the allocation values for the closure.
* Using a hash map to store the preferred offsets for the computations.
* Using a hash map to store the copy destinations for the copy instructions.
* Using a hash map to store the copy sources for the copy instructions.
* Using a hash map to store the required memory assignments for the instructions.

PiperOrigin-RevId: 643097604